refactor: use common helper for getFormattedTimeToLive
