+++ src/Pages/GlobalConfigurations/Authorization/libUtils.ts
@@ -3,9 +3,16 @@ import { importComponentFromFELibrary } from '../../../components/common'
 import { User } from './types'
 
 // Adding here to avoid circular import dependency
-// These two are being reused across with default values
+// These are being reused across with default values
+
 export const getDefaultUserStatusAndTimeout: () => Pick<User, 'userStatus' | 'timeToLive'> =
     importComponentFromFELibrary('getDefaultUserStatusAndTimeout', () => ({}), 'function')
 
 export const getDefaultStatusAndTimeout: () => Pick<UserRoleGroup, 'status' | 'timeToLive'> =
     importComponentFromFELibrary('getDefaultStatusAndTimeout', () => ({}), 'function')
+
+export const getFormattedTimeToLive: (timeToLive: string) => string = importComponentFromFELibrary(
+    'getFormattedTimeToLive',
+    () => ({}),
+    'function',
+)
